+++ tools.h
@@ -13,7 +13,7 @@ int file_exists (char * fileName);
 void set_mac();
 int mkdir_r(std::string, mode_t);
 double ts2double(unsigned int sec, unsigned int usec);
-long GetFileSize(std::string filename);
+unsigned long long GetFileSize(std::string filename);
 bool FileExists(char *strFilename);
 void ntoa(char *res, unsigned int addr);
 string escapeshellR(string &);
